just overflow

+++ linear-assignment.h
@@ -13,8 +13,8 @@
  * assignments (-1 for unassigned, which can happen only if column_count !=
  * row_count).
  */
-void compute_assignment(int column_count, int row_count, int *cost,
-			int *column2row, int *row2column);
+void compute_assignment(ssize_t column_count, ssize_t row_count, ssize_t *cost,
+			ssize_t *column2row, ssize_t *row2column);
 
 /* The maximal cost in the cost matrix (to prevent integer overflows). */
 #define COST_MAX (1<<16)
